fix: cart not found error when cancelling or completing payment (#18676)
CXSPA-6537

+++ integration-libs/opf/base/components/opf-quick-buy/apple-pay/apple-pay.service.ts
@@ -140,18 +140,24 @@ export class ApplePayService {
       }),
       take(1),
       catchError((error) => {
-        this.cartHandlerService.loadCartAfterSingleProductTransaction(
-          this.transactionDetails
-        );
-        return throwError(() => error);
+        return this.cartHandlerService
+          .loadCartAfterSingleProductTransaction(this.transactionDetails)
+          .pipe(switchMap(() => throwError(() => error)));
       }),
       finalize(() => {
-        this.cartHandlerService.deleteUserAddresses([
-          ...this.transactionDetails.addressIds,
-        ]);
+        this.deleteUserAddresses();
         this.paymentInProgress = false;
       })
     );
+  }
+
+  protected deleteUserAddresses() {
+    if (this.transactionDetails.addressIds.length) {
+      this.cartHandlerService.deleteUserAddresses([
+        ...this.transactionDetails.addressIds,
+      ]);
+      this.transactionDetails.addressIds = [];
+    }
   }
 
   private handleValidation(
@@ -360,14 +366,17 @@ export class ApplePayService {
     const result: ApplePayJS.ApplePayPaymentAuthorizationResult = {
       status: this.applePaySession.statusSuccess,
     };
-
+    let orderSuccess: boolean;
     return this.placeOrderAfterPayment(event.payment).pipe(
-      map((success) => {
-        this.cartHandlerService.loadCartAfterSingleProductTransaction(
+      switchMap((success) => {
+        orderSuccess = success;
+        return this.cartHandlerService.loadCartAfterSingleProductTransaction(
           this.transactionDetails,
           success
         );
-        return success
+      }),
+      map(() => {
+        return orderSuccess
           ? result
           : { ...result, status: this.applePaySession.statusFailure };
       }),
